upload local auto_update_settings

+++ admin/partials/webchangedetector-admin-display.php
@@ -215,8 +215,8 @@ if ( ! function_exists( 'wcd_webchangedetector_init' ) ) {
 
 		// Moving local auto update settings to the api.
 		$local_auto_update_settings = get_option( WCD_AUTO_UPDATE_SETTINGS );
-		if ( $local_auto_update_settings ) {
-			$wcd->update_website_details( $local_auto_update_settings );
+		if ( $local_auto_update_settings && empty($wcd->website_details['auto_update_settings'] ) ) {
+			$wcd->update_website_details( ['auto_update_settings' => $local_auto_update_settings] );
 			delete_option( WCD_AUTO_UPDATE_SETTINGS );
 		}
